Implementation and test of IncrementalEvaluationMatrix
The class IncrementalEvaluationMatrix is implemented to reduce the computational inefficiency of recalculating connections in each evaluation. The iterative matrix solution was temporary unimplemented to make this change. The solution was tested and its identical to previous implementations (but can be really faster with small changes to implement in futur commits).